Start of keyloggeer. Hook and callback function setup.
Got the start of the keylogger setup. Calls the SetWindowsHookEx add a
WH_KEYBOARD_LL hook to the chain. The callback function 'HookCallBack'
checks for a key press, sends a popup box to the user and unhooks
itself. The last thing the call back function does is send a WM_QUIT
message using PostQuitMessage(). This causes the program to exit the
message loop and stops the program.