Mapped out the virtual keys for the keyboard.
Used a map container to map the virtual id to their keys. Used virtual
keys over scan code since virtual keys are device independent. Added a
debug mode that prints the virtual code and scan code of a key to a
prompt window. Moved the unhook function out of the callback function
and outside of the message loop. Program will now check for any errors
when calling the unhook function and print the error to console.